Remove templates and rename game controller to games controller
